Add files via upload
A first version of Data Minting Algorithm.

The test file is mario.gif at the size of 1ko. The algorithm generate a minted 2ko file which is bigger.

The goal is to optimize the Traveler Salesman Problem, and the encode to such an extent can mint a file to a smaller size.